Cleanup json formatters.
