+++ manager/src/main/scala/io/hydrosphere/serving/manager/controller/model_source/ModelSourceController.scala
@@ -3,14 +3,15 @@ package io.hydrosphere.serving.manager.controller.model_source
 import javax.ws.rs.Path
 
 import akka.http.scaladsl.server.Directives._
-import io.hydrosphere.serving.manager.model.{ManagerJsonSupport, ModelSourceConfigAux}
-import io.hydrosphere.serving.manager.service.{SourceManagementService}
+import io.hydrosphere.serving.manager.model.ModelSourceConfigAux
+import io.hydrosphere.serving.manager.service.SourceManagementService
+import io.hydrosphere.serving.manager.model.CommonJsonSupport._
 import io.swagger.annotations._
 
 
 @Path("/api/v1/modelSource")
 @Api(produces = "application/json", tags = Array("Model Sources"))
-class ModelSourceController(sourceService: SourceManagementService) extends ManagerJsonSupport {
+class ModelSourceController(sourceService: SourceManagementService) {
   @Path("/local")
   @ApiOperation(value = "Add local model source", notes = "Add local model source", nickname = "addLocalSource", httpMethod = "POST")
   @ApiImplicitParams(Array(
