+++ manager/src/main/scala/io/hydrosphere/serving/manager/repository/db/ModelRepositoryImpl.scala
@@ -4,20 +4,18 @@ import java.time.LocalDateTime
 
 import io.hydrosphere.serving.contract.model_contract.ModelContract
 import io.hydrosphere.serving.manager.db.Tables
-import io.hydrosphere.serving.manager.model.{ManagerJsonSupport, Model}
+import io.hydrosphere.serving.manager.model.Model
 import io.hydrosphere.serving.manager.model.api.ModelType
 import io.hydrosphere.serving.manager.repository.ModelRepository
+import io.hydrosphere.serving.manager.model.CommonJsonSupport._
 import org.apache.logging.log4j.scala.Logging
 
 import scala.concurrent.{ExecutionContext, Future}
 
-/**
-  *
-  */
 class ModelRepositoryImpl(
   implicit executionContext: ExecutionContext,
   databaseService: DatabaseService
-) extends ModelRepository with Logging with ManagerJsonSupport {
+) extends ModelRepository with Logging {
 
   import databaseService._
   import databaseService.driver.api._
@@ -117,7 +115,7 @@ class ModelRepositoryImpl(
     ).map(mapFromDb)
 }
 
-object ModelRepositoryImpl extends ManagerJsonSupport {
+object ModelRepositoryImpl {
   def mapFromDb(model: Option[Tables.Model#TableElementType]): Option[Model] =
     model.map(mapFromDb)
